fundora-mtcnn-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
mtcnn==0.1.1
tensorflow==2.13.0
keras==2.13.0